Extracting TinyXML2 as external dependency
